added crud operations and schemas
