+++ app.js
@@ -1,8 +1,6 @@
 const express = require("express");
 const logger = require("morgan");
 const cors = require("cors");
-const mongoose = require("mongoose");
-require("dotenv").config();
 
 const contactsRouter = require("./routes/api/contacts");
 
@@ -15,19 +13,6 @@ app.use(cors());
 app.use(express.json());
 
 app.use("/api/contacts", contactsRouter);
-
-const { DB_HOST, PORT = 3000 } = process.env;
-
-mongoose
-  .connect(DB_HOST)
-  .then(() => {
-    console.log("Database connection successful");
-  })
-  .then(() => app.listen(PORT))
-  .catch((error) => {
-    console.log(error.message);
-    process.exit(1);
-  });
 
 app.use((req, res) => {
   res.status(404).json({ message: "Not found" });
