+++ controllers/contacts/getById.js
@@ -1,13 +1,19 @@
-const contacts = require("../../models/contact");
+const { Contact } = require("../../models/contact");
 const { createError } = require("../../helpers");
 
 const getById = async (req, res, next) => {
   const { id } = req.params;
-  const result = await contacts.getById(id);
+  const result = await Contact.findById(id);
   if (!result) {
     throw createError(404);
   }
-  res.json(result);
+  res.json({
+    status: "success",
+    code: 200,
+    data: {
+      result,
+    },
+  });
 };
 
 module.exports = getById;
