+++ controllers/contacts/listContacts.js
@@ -1,4 +1,4 @@
-const { Contact } = require("../../models");
+const { Contact } = require("../../models/contact");
 
 const listContacts = async (req, res) => {
   const result = await Contact.find({});
