+++ controllers/contacts/removeContact.js
@@ -1,9 +1,9 @@
 const { createError } = require("../../helpers");
-const contacts = require("../../models/contact");
+const { Contact } = require("../../models/contact");
 
 const removeContact = async (req, res, next) => {
   const { id } = req.params;
-  const result = await contacts.removeContact(id);
+  const result = await Contact.findByIdAndRemove(id);
   if (!result) {
     throw createError(404);
   }
