+++ models/contact.js
@@ -1,22 +1,51 @@
 const { Schema, model } = require("mongoose");
 
-const contactSchema = new Schema({
-  name: {
-    type: String,
-    required: [true, "Set name for contact"],
+const Joi = require("joi");
+
+const contactSchema = new Schema(
+  {
+    name: {
+      type: String,
+      required: [true, "Set name for contact"],
+    },
+    email: {
+      type: String,
+      required: true,
+      unique: true,
+    },
+    phone: {
+      type: String,
+      required: true,
+    },
+    favorite: {
+      type: Boolean,
+      default: false,
+    },
   },
-  email: {
-    type: String,
-  },
-  phone: {
-    type: String,
-  },
-  favorite: {
-    type: Boolean,
-    default: false,
-  },
-});
+  { versionKey: false, timestamps: true }
+);
 
 const Contact = model("contact", contactSchema);
 
-module.exports = Contact;
+const contactsAddSchema = Joi.object({
+  name: Joi.string().required(),
+  email: Joi.string()
+    .email({ minDomainSegments: 2, tlds: { allow: ["com", "net"] } })
+    .required(),
+  phone: Joi.number().required(),
+  favorite: Joi.bool(),
+});
+
+const updateFavoriteSchema = Joi.object({
+  favorite: Joi.bool().required(),
+});
+
+const schemas = {
+  contactsAddSchema,
+  updateFavoriteSchema,
+};
+
+module.exports = {
+  Contact,
+  schemas,
+};
