--- models/index.js
@@ -1,5 +1,0 @@
-const Contact = require("./contact");
-
-module.exports = {
-  Contact,
-};
