--- schemas/contacts.js
@@ -1,13 +1,0 @@
-const Joi = require("joi");
-
-const contactsAddSchema = Joi.object({
-  name: Joi.string().required(),
-  email: Joi.string()
-    .email({ minDomainSegments: 2, tlds: { allow: ["com", "net"] } })
-    .required(),
-  phone: Joi.number().required(),
-});
-
-module.exports = {
-  contactsAddSchema,
-};
