+++ server.js
@@ -1,5 +1,16 @@
-// const app = require('./app')
+const app = require("./app");
+const mongoose = require("mongoose");
+require("dotenv").config();
 
-// app.listen(3000, () => {
-//   console.log("Server running. Use our API on port: 3000")
-// })
+const { DB_HOST, PORT = 3000 } = process.env;
+
+mongoose
+  .connect(DB_HOST)
+  .then(() => {
+    console.log("Database connection successful");
+  })
+  .then(() => app.listen(PORT))
+  .catch((error) => {
+    console.log(error.message);
+    process.exit(1);
+  });
